Introduce a script_layout_interface crate and move RestyleDamage to it.
