+++ components/layout/Cargo.toml
@@ -31,6 +31,7 @@ profile_traits = {path = "../profile_traits"}
 range = {path = "../range"}
 rustc-serialize = "0.3"
 script = {path = "../script"}
+script_layout_interface = {path = "../script_layout_interface"}
 script_traits = {path = "../script_traits"}
 selectors = {version = "0.6", features = ["heap_size"]}
 serde_json = "0.7"
