+++ components/layout/table_cell.rs
@@ -16,9 +16,9 @@ use flow::{self, Flow, FlowClass, OpaqueFlow};
 use fragment::{Fragment, FragmentBorderBoxIterator, Overflow};
 use gfx::display_list::StackingContext;
 use gfx_traits::StackingContextId;
-use incremental::REFLOW;
 use layout_debug;
 use model::MaybeAuto;
+use script_layout_interface::restyle_damage::REFLOW;
 use std::fmt;
 use std::sync::Arc;
 use style::computed_values::{border_collapse, border_top_style, vertical_align};
